+++ dalesbred/src/main/java/org/dalesbred/instantiation/CoercionInstantiator.java
@@ -35,13 +35,8 @@ final class CoercionInstantiator<T> implements Instantiator<T> {
     @NotNull
     private final TypeConversion<Object, ? extends T> coercion;
 
-    @Nullable
-    private final InstantiationListener instantiationListener;
-
-    CoercionInstantiator(@NotNull TypeConversion<Object, ? extends T> coercion,
-                         @Nullable InstantiationListener instantiationListener) {
+    CoercionInstantiator(@NotNull TypeConversion<Object, ? extends T> coercion) {
         this.coercion = requireNonNull(coercion);
-        this.instantiationListener = instantiationListener;
     }
 
     @Nullable
@@ -51,10 +46,7 @@ final class CoercionInstantiator<T> implements Instantiator<T> {
 
         Object value = arguments.getValues().get(0);
         if (value != null) {
-            T result = coercion.convert(value);
-            if (instantiationListener != null)
-                instantiationListener.onInstantiation(result);
-            return result;
+            return coercion.convert(value);
         } else
             return null;
     }