+++ dalesbred/src/main/java/org/dalesbred/instantiation/InstantiatorRegistry.java
@@ -30,11 +30,6 @@ import org.jetbrains.annotations.NotNull;
 public interface InstantiatorRegistry {
 
     /**
-     * Adds a new InstantiationListener, which gets notified whenever one of the instantiators creates an object.
-     */
-    void addInstantiationListener(@NotNull InstantiationListener instantiationListener);
-
-    /**
      * Registers an instantiator to be used when instances of {@code cl} are required.
      */
     <T> void registerInstantiator(@NotNull Class<T> cl, @NotNull Instantiator<T> instantiator);